move FactBox logic into own file
